backend/requirements.txt
asgiref==3.2.7
astroid==2.4.1
bson==0.5.8
censys==0.0.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
dataclasses==0.6
Django==3.0.6
djangorestframework==3.11.0
djongo==1.3.2
# dnspython==2.0.0
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.8
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.8.0
netaddr==0.7.19
oauth2client==3.0.0
protobuf==3.11.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pymongo==3.10.1
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
-e git+https://github.com/pawlaczyk/sarenka.git@b217f3ffb5f4eff2d52cbba6a30da3f05337311a#egg=sarenka&subdirectory=backend
shodan==1.21.3
six==1.14.0
sqlparse==0.2.4
toml==0.10.1
typing==3.7.4.1
urllib3==1.25.8
wrapt==1.12.1
XlsxWriter==1.2.7
